Lots of changes!
Got rid of the cmdserver because we are now an ipython application.

Created a PipeLineTasks list. Now all analysis tasks (functions )will be
placed in this list and called in a safe way (with error handling). The
tasks cascade on after another and the output from on is the unput
to the next. The input and output should always be an HDUlist data type
from the astropy.io.fits library. This HDUlist is then written to a file
in the sub directory reduced_images. Which will be created if it does not
exist.

We also now have an ImageDataList which gives a running tally of some usefull
image data in a list of dictionary objects. Data is added to the dictionary
with the saveImageInfo2List function. This is still pretty primitive and could
easily cause trouble especially if an image wasn't finished processing before
a new image came through.

TODO:
Create timeouts for each pipeline task so as to make sure the image finishes
processing before the next image comes through. Create a series of pipeline
management functions perhaps wrapping them in a class so the user won't be
able to directly access the PipeLineTasks list or the ImageDataList.